+++ src/lib/primevue/components/checkbox/sizes/index.ts
@@ -1,0 +1,31 @@
+const sizes = {
+  root: {
+    borderRadius: '{border.radius.sm}',
+    width: '1.25rem',
+    height: '1.25rem',
+    sm: {
+      width: '1rem',
+      height: '1rem'
+    },
+    lg: {
+      width: '1.5rem',
+      height: '1.5rem'
+    },
+    focusRing: {
+      width: '{focus.ring.width}',
+      style: '{focus.ring.style}',
+      offset: '{focus.ring.offset}',
+    }
+  },
+  icon: {
+    size: '0.875rem',
+    sm: {
+      size: '0.75rem'
+    },
+    lg: {
+      size: '1rem'
+    }
+  }
+}
+
+export default sizes