+++ src/lib/primevue/components/index.ts
@@ -5,6 +5,7 @@ import TooltipScheme from './tooltip';
 import MenubarScheme from './menubar';
 import ToolbarScheme from './toolbar';
 import ChipScheme from "./chip";
+import CheckboxScheme from "./checkbox";
 
 export {
   AvatarScheme,
@@ -14,4 +15,5 @@ export {
   MenubarScheme,
   ToolbarScheme,
   ChipScheme,
+  CheckboxScheme,
 }